Added Crafting and Menu Swapping
Tackled the crafting menu next. Needed a way to swap between the menus so I actually stole some of the logic (now in Tabs.cs) from the lethal company project as a way to increment and decrement between the menu pages. All I do is make them visible or not and everything works nicely.