Successful metadata usage
Instead of making another custom signal just for removing from the list, I added a parameter for the amount added and just add 1 for left click and -1 for right click.

"But then shouldnt the signal be called something a little different to better show what it's doi-" shutup. Im adding a negative number. its fine