Sorting
Oh gosh things go by quickly when you get the right setup from the beginning. Just had to make 2 buttons and figure out the sort method. One detail to point out is that i forgot the values are Variants, so when actually sorting, LINQ didnt know what to do. Had to change the metadata to integers