Clear Button
Had an obnoxious time with this one. Seemed like I was doing everything right, using the Tabs Add signal as an example. Turns out visual studio was not showing me the correct name for the custom signal. It was showing me the Name with EventHandler at the end, but when emitting the signal you need to EXCLUDE EventHandler. **The more you know**